+++ actors/migration/nv14/test/parallel_migration_test.go
@@ -2,30 +2,13 @@ package test_test
 
 import (
 	"context"
-	"fmt"
 	"testing"
 
-	"github.com/filecoin-project/go-address"
-	"github.com/filecoin-project/go-bitfield"
 	"github.com/filecoin-project/go-state-types/abi"
-	"github.com/filecoin-project/go-state-types/big"
-	"github.com/filecoin-project/go-state-types/dline"
-	"github.com/filecoin-project/go-state-types/rt"
-	miner0 "github.com/filecoin-project/specs-actors/actors/builtin/miner"
 	ipld2 "github.com/filecoin-project/specs-actors/v2/support/ipld"
-	builtin5 "github.com/filecoin-project/specs-actors/v5/actors/builtin"
-	miner5 "github.com/filecoin-project/specs-actors/v5/actors/builtin/miner"
-	power5 "github.com/filecoin-project/specs-actors/v5/actors/builtin/power"
 	vm5 "github.com/filecoin-project/specs-actors/v5/support/vm"
-	"github.com/filecoin-project/specs-actors/v6/actors/builtin"
-	"github.com/filecoin-project/specs-actors/v6/actors/builtin/exported"
-	"github.com/filecoin-project/specs-actors/v6/actors/builtin/miner"
-	"github.com/filecoin-project/specs-actors/v6/actors/builtin/power"
 	"github.com/filecoin-project/specs-actors/v6/actors/migration/nv14"
-	"github.com/filecoin-project/specs-actors/v6/actors/runtime/proof"
 	adt5 "github.com/filecoin-project/specs-actors/v6/actors/util/adt"
-	tutil "github.com/filecoin-project/specs-actors/v6/support/testing"
-	"github.com/filecoin-project/specs-actors/v6/support/vm"
 	cbor "github.com/ipfs/go-ipld-cbor"
 
 	"github.com/ipfs/go-cid"
@@ -63,220 +46,4 @@ func TestParallelMigrationCalls(t *testing.T) {
 	require.NoError(t, grp.Wait())
 	assert.Equal(t, endRootSerial, endRootParallel1)
 	assert.Equal(t, endRootParallel1, endRootParallel2)
-}
-
-func TestEarlyTerminationInCronBeforeAndAfterMigration(t *testing.T) {
-
-	ctx := context.Background()
-	bs := ipld2.NewSyncBlockStoreInMemory()
-	v := vm5.NewVMWithSingletons(ctx, t, bs)
-
-	sealProof := abi.RegisteredSealProof_StackedDrg32GiBV1_1
-	wPoStProof, err := sealProof.RegisteredWindowPoStProof()
-	require.NoError(t, err)
-
-	require.NoError(t, err)
-	addrs := vm5.CreateAccounts(ctx, t, v, 1, big.Mul(big.NewInt(10_000), builtin.TokenPrecision), 93837778)
-	owner, worker := addrs[0], addrs[0]
-	minerAddrs := CreateMinerV5(t, v, owner, worker, wPoStProof, big.Mul(big.NewInt(10_000), vm.FIL))
-
-	// Pre commit two sectors
-	v, err = v.WithEpoch(200)
-	require.NoError(t, err)
-	sectorNumber := abi.SectorNumber(100)
-	PreCommitSectorsV5(t, v, 2, 1, worker, minerAddrs.IDAddress, sealProof, sectorNumber, true)
-	balances := vm5.GetMinerBalances(t, v, minerAddrs.IDAddress)
-	assert.True(t, balances.PreCommitDeposit.GreaterThan(big.Zero()))
-	proveTime := v.GetEpoch() + miner.MaxProveCommitDuration[sealProof]
-	v, _ = vm5.AdvanceByDeadlineTillEpoch(t, v, minerAddrs.IDAddress, proveTime)
-
-	// Prove commit two sectors
-	v, err = v.WithEpoch(proveTime)
-	require.NoError(t, err)
-	proveCommitParams := miner.ProveCommitSectorParams{
-		SectorNumber: sectorNumber,
-	}
-	vm5.ApplyOk(t, v, addrs[0], minerAddrs.RobustAddress, big.Zero(), builtin.MethodsMiner.ProveCommitSector, &proveCommitParams)
-	proveCommitParams.SectorNumber = sectorNumber + 1
-	vm5.ApplyOk(t, v, addrs[0], minerAddrs.RobustAddress, big.Zero(), builtin.MethodsMiner.ProveCommitSector, &proveCommitParams)
-	// In the same epoch, trigger cron to validate prove commit
-	vm5.ApplyOk(t, v, builtin.SystemActorAddr, builtin.CronActorAddr, big.Zero(), builtin.MethodsCron.EpochTick, nil)
-
-	// PoSt
-	sectorSize, err := sealProof.SectorSize()
-	require.NoError(t, err)
-	dlInfo, pIdx, v := vm5.AdvanceTillProvingDeadline(t, v, minerAddrs.IDAddress, sectorNumber)
-	var minerState miner.State
-	err = v.GetState(minerAddrs.IDAddress, &minerState)
-	require.NoError(t, err)
-	sector1, found, err := minerState.GetSector(v.Store(), sectorNumber)
-	require.NoError(t, err)
-	require.True(t, found)
-	sector2, found, err := minerState.GetSector(v.Store(), sectorNumber+1)
-	require.NoError(t, err)
-	require.True(t, found)
-	v, err = v.WithEpoch(v.GetEpoch())
-	require.NoError(t, err)
-	partitions := []miner.PoStPartition{{
-		Index:   pIdx,
-		Skipped: bitfield.New(),
-	}}
-	sectorPower := miner.PowerForSector(sectorSize, sector1)
-	sectorPower = sectorPower.Add(miner.PowerForSector(sectorSize, sector2))
-	SubmitWindowPoStV5(t, v, worker, minerAddrs.IDAddress, dlInfo, partitions, sectorPower)
-	v, err = v.WithEpoch(dlInfo.Last()) // run cron on deadline end to process post
-	require.NoError(t, err)
-	vm5.ApplyOk(t, v, builtin.SystemActorAddr, builtin.CronActorAddr, big.Zero(), builtin.MethodsCron.EpochTick, nil)
-
-	// Fault sector 100 by skipping in post in the next proving period
-	dlInfo, pIdx, v = vm5.AdvanceTillProvingDeadline(t, v, minerAddrs.IDAddress, sectorNumber)
-	//	fmt.Printf("after dlInfo.Index: %d, dlInfo.Open: %d, dlInfo.Close: %d, pIdx: %d, curr epoch: %d\n", dlInfo.Index, dlInfo.Open, dlInfo.Close, pIdx, v.GetEpoch())
-	partitions = []miner.PoStPartition{{
-		Index:   pIdx,
-		Skipped: bitfield.NewFromSet([]uint64{uint64(sectorNumber)}),
-	}}
-	v, err = v.WithEpoch(dlInfo.Last()) // run cron on deadline end to fault
-	require.NoError(t, err)
-	SubmitWindowPoStV5(t, v, worker, minerAddrs.IDAddress, dlInfo, partitions, miner.PowerForSector(sectorSize, sector1).Neg())
-	vm5.ApplyOk(t, v, builtin.SystemActorAddr, builtin.CronActorAddr, big.Zero(), builtin.MethodsCron.EpochTick, nil)
-
-	// Migrate
-	log := nv14.TestLogger{TB: t}
-	adtStore := adt5.WrapStore(ctx, cbor.NewCborStore(bs))
-	startRoot := v.StateRoot()
-	nextRoot, err := nv14.MigrateStateTree(ctx, adtStore, startRoot, abi.ChainEpoch(0), nv14.Config{MaxWorkers: 1}, log, nv14.NewMemMigrationCache())
-	require.NoError(t, err)
-
-	lookup := map[cid.Cid]rt.VMActor{}
-	for _, ba := range exported.BuiltinActors() {
-		lookup[ba.Code()] = ba
-	}
-
-	v6, err := vm.NewVMAtEpoch(ctx, lookup, v.Store(), nextRoot, v.GetEpoch()+1)
-	require.NoError(t, err)
-
-	// Fault sector 101
-	d, p := vm.SectorDeadline(t, v6, minerAddrs.IDAddress, sectorNumber+1)
-	v6, _ = vm.AdvanceByDeadlineTillIndex(t, v6, minerAddrs.IDAddress, d+2) // move out of deadline so fault can go through
-	dfParams := &miner5.DeclareFaultsParams{
-		Faults: []miner0.FaultDeclaration{
-			{
-				Deadline:  d,
-				Partition: p,
-				Sectors:   bitfield.NewFromSet([]uint64{uint64(sectorNumber + 1)}),
-			},
-		},
-	}
-	vm.ApplyOk(t, v6, worker, minerAddrs.IDAddress, big.Zero(), builtin5.MethodsMiner.DeclareFaults, dfParams)
-
-}
-
-func CreateMinerV5(t *testing.T, v *vm5.VM, owner, worker address.Address, wPoStProof abi.RegisteredPoStProof, balance abi.TokenAmount) *power.CreateMinerReturn {
-	params := power5.CreateMinerParams{
-		Owner:               owner,
-		Worker:              worker,
-		WindowPoStProofType: wPoStProof,
-		Peer:                abi.PeerID("not really a peer id"),
-	}
-	ret := vm5.ApplyOk(t, v, worker, builtin.StoragePowerActorAddr, balance, builtin.MethodsPower.CreateMiner, &params)
-	minerAddrs, ok := ret.(*power.CreateMinerReturn)
-	require.True(t, ok)
-	return minerAddrs
-}
-
-func PreCommitSectorsV5(t *testing.T, v *vm5.VM, count, batchSize int, worker, mAddr address.Address, sealProof abi.RegisteredSealProof,
-	sectorNumberBase abi.SectorNumber, expectCronEnrollment bool) []*miner.SectorPreCommitOnChainInfo {
-	invocsCommon := []vm5.ExpectInvocation{
-		{To: builtin.RewardActorAddr, Method: builtin.MethodsReward.ThisEpochReward},
-		{To: builtin.StoragePowerActorAddr, Method: builtin.MethodsPower.CurrentTotalPower},
-	}
-	invocFirst := vm5.ExpectInvocation{To: builtin.StoragePowerActorAddr, Method: builtin.MethodsPower.EnrollCronEvent}
-
-	sectorIndex := 0
-	for sectorIndex < count {
-		msgSectorIndexStart := sectorIndex
-		invocs := invocsCommon
-
-		// Prepare message.
-		params := miner5.PreCommitSectorBatchParams{Sectors: make([]miner0.SectorPreCommitInfo, batchSize)}
-		for j := 0; j < batchSize && sectorIndex < count; j++ {
-			sectorNumber := sectorNumberBase + abi.SectorNumber(sectorIndex)
-			sealedCid := tutil.MakeCID(fmt.Sprintf("%d", sectorNumber), &miner.SealedCIDPrefix)
-			params.Sectors[j] = miner0.SectorPreCommitInfo{
-				SealProof:     sealProof,
-				SectorNumber:  sectorNumber,
-				SealedCID:     sealedCid,
-				SealRandEpoch: v.GetEpoch() - 1,
-				DealIDs:       nil,
-				Expiration:    v.GetEpoch() + miner.MinSectorExpiration + miner.MaxProveCommitDuration[sealProof] + 100,
-			}
-			sectorIndex++
-		}
-		if sectorIndex == count && sectorIndex%batchSize != 0 {
-			// Trim the last, partial batch.
-			params.Sectors = params.Sectors[:sectorIndex%batchSize]
-		}
-
-		// Finalize invocation expectation list
-		if len(params.Sectors) > 1 {
-			aggFee := miner.AggregatePreCommitNetworkFee(len(params.Sectors), big.Zero())
-			invocs = append(invocs, vm5.ExpectInvocation{To: builtin.BurntFundsActorAddr, Method: builtin.MethodSend, Value: &aggFee})
-		}
-		if expectCronEnrollment && msgSectorIndexStart == 0 {
-			invocs = append(invocs, invocFirst)
-		}
-		vm5.ApplyOk(t, v, worker, mAddr, big.Zero(), builtin.MethodsMiner.PreCommitSectorBatch, &params)
-		vm5.ExpectInvocation{
-			To:             mAddr,
-			Method:         builtin.MethodsMiner.PreCommitSectorBatch,
-			Params:         vm5.ExpectObject(&params),
-			SubInvocations: invocs,
-		}.Matches(t, v.LastInvocation())
-	}
-
-	// Extract chain state.
-	var minerState miner.State
-	err := v.GetState(mAddr, &minerState)
-	require.NoError(t, err)
-
-	precommits := make([]*miner.SectorPreCommitOnChainInfo, count)
-	for i := 0; i < count; i++ {
-		precommit, found, err := minerState.GetPrecommittedSector(v.Store(), sectorNumberBase+abi.SectorNumber(i))
-		require.NoError(t, err)
-		require.True(t, found)
-		precommits[i] = precommit
-	}
-	return precommits
-}
-
-func SubmitWindowPoStV5(t *testing.T, v *vm5.VM, worker, actor address.Address, dlInfo *dline.Info, partitions []miner.PoStPartition,
-	newPower miner.PowerPair) {
-	submitParams := miner.SubmitWindowedPoStParams{
-		Deadline:   dlInfo.Index,
-		Partitions: partitions,
-		Proofs: []proof.PoStProof{{
-			PoStProof: abi.RegisteredPoStProof_StackedDrgWindow32GiBV1,
-		}},
-		ChainCommitEpoch: dlInfo.Challenge,
-		ChainCommitRand:  []byte(vm.RandString),
-	}
-	vm5.ApplyOk(t, v, worker, actor, big.Zero(), builtin.MethodsMiner.SubmitWindowedPoSt, &submitParams)
-
-	updatePowerParams := &power.UpdateClaimedPowerParams{
-		RawByteDelta:         newPower.Raw,
-		QualityAdjustedDelta: newPower.QA,
-	}
-
-	vm5.ExpectInvocation{
-		To:     actor,
-		Method: builtin.MethodsMiner.SubmitWindowedPoSt,
-		Params: vm5.ExpectObject(&submitParams),
-		SubInvocations: []vm5.ExpectInvocation{
-			{
-				To:     builtin.StoragePowerActorAddr,
-				Method: builtin.MethodsPower.UpdateClaimedPower,
-				Params: vm5.ExpectObject(updatePowerParams),
-			},
-		},
-	}.Matches(t, v.LastInvocation())
 }
